new class camera + updated position of the player to float instead of long

+++ code/src/java/futuremctests/Joueur.java
@@ -1,7 +1,7 @@
 package futuremctests;
 
 public class Joueur {
-    public long x;
-    public long y;
+    public float x;
+    public float y;
 
 }
